gbc.ml.document.classifier.extra_trees/requirements.txt
# pip install setuptools --upgrade --ignore-installed
# python -m spacy download es_core_news_md/en_core_web_md

elasticsearch==7.5.1
certifi==2019.11.28
binaryornot==0.4.4
numpy==1.18.1
scikit-learn==0.22.1
flask==1.1.1
flasgger==0.9.4
Flask-Cors==3.0.8
Flask-JWT-Extended==3.24.1
flask-restplus==0.13.0
boto3==1.11.11
opentracing==2.3.0
opentracing-utils==0.18.1
jaeger-client==4.3.0
spacy==2.2.3
requests==2.22.0
httpx==0.11.1
tensorflow==2.1.0
Pillow==7.0.0
opencv-python==4.2.0.32
imutils==0.5.3
scipy==1.4.1
healthcheck==1.3.3
python-Levenshtein==0.12.0
PyDictionary==1.5.2
beautifulsoup4==4.8.2
Werkzeug==0.16.1
PyYAML==5.3
